requirements.txt
packaging
Django
pyschopg2
pytz
google-cloud-storage
google-cloud-firestore
